Refactor verifier to work with AbstractRuleSet
Will allow use with rules that extend AI rules now.
